requirements.txt
langchain
python-dotenv # to configure environment variables
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
sentence-transformers
chromadb
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette